+++ cmd/bnsd/app/gov_handler.go
@@ -3,6 +3,7 @@ package bnsd
 import (
 	"github.com/iov-one/weave"
 	"github.com/iov-one/weave/app"
+	"github.com/iov-one/weave/cmd/bnsd/x/termdeposit"
 	"github.com/iov-one/weave/errors"
 	"github.com/iov-one/weave/migration"
 	"github.com/iov-one/weave/x/batch"
@@ -11,7 +12,7 @@ import (
 	"github.com/iov-one/weave/x/escrow"
 	"github.com/iov-one/weave/x/gov"
 	"github.com/iov-one/weave/x/msgfee"
-	txfee "github.com/iov-one/weave/x/txfee"
+	"github.com/iov-one/weave/x/txfee"
 	"github.com/iov-one/weave/x/utils"
 	"github.com/iov-one/weave/x/validators"
 )
@@ -43,6 +44,7 @@ func proposalOptionsExecutor(ctrl cash.Controller) gov.Executor {
 	gov.RegisterBasicProposalRouters(r, auth)
 	msgfee.RegisterRoutes(r, auth)
 	txfee.RegisterRoutes(r, auth)
+	termdeposit.RegisterRoutes(r, auth, ctrl)
 
 	// We must wrap with batch middleware so it can process ExecuteProposalBatchMsg.
 	// We add ActionTagger here, so the messages executed as a result of a governance vote also get properly tagged.