+++ cmd/bnsd/x/termdeposit/doc.go
@@ -1,0 +1,9 @@
+/*
+Package termdeposit implements an intrest-bearing contracts. Anyone can sign a
+contract by depositing funds.
+
+This is a minimal implementation of term deposit functionality. Each deposit
+interest is computed offchain and transferred inteprendetly from this
+extension.
+*/
+package termdeposit